server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2020.12.5        # via requests
chardet==4.0.0            # via requests
click==7.1.2              # via flask
flask==1.1.2              # via -r requirements.in
future==0.18.2            # via python-twitter
idna==2.10                # via requests
itsdangerous==1.1.0       # via flask
jinja2==2.11.2            # via flask
markupsafe==1.1.1         # via jinja2
oauthlib==3.1.0           # via requests-oauthlib
pillow==8.0.1             # via -r requirements.in
python-dotenv==0.15.0     # via -r requirements.in
python-twitter==3.5       # via -r requirements.in
requests-oauthlib==1.3.0  # via python-twitter
requests==2.25.1          # via -r requirements.in, python-twitter, requests-oauthlib
urllib3==1.26.2           # via requests
werkzeug==1.0.1           # via flask
